Increase encoder resolution to 10 bits

+++ pandora_xmega_hardware_interface/include/pandora_xmega_hardware_interface/encoder_sensor.h
@@ -54,7 +54,7 @@ class EncoderSensor : virtual public SensorBase
 
  public:
   
-  int degrees;
+  float degrees;
 };
 
 }  // namespace xmega
