+++ pandora_xmega_hardware_interface/include/pandora_xmega_hardware_interface/xmega_serial_interface.h
@@ -109,7 +109,7 @@ class XmegaSerialInterface : private boost::noncopyable
     return rangeSensors_.sensors;
   }
 
-  inline int getEncoderDegrees() const
+  inline float getEncoderDegrees() const
   {
     return encoderSensor_.degrees;
   }
